# Improve slice.binary_search_by()'s best-case performance to O(1)