--- a/library/core/src/slice/mod.rs
+++ b/library/core/src/slice/mod.rs
@@ -2167,7 +2167,11 @@ impl<T> [T] {
             // - `mid >= 0`: by definition
             // - `mid < size`: `mid = size / 2 + size / 4 + size / 8 ...`
             let cmp = f(unsafe { s.get_unchecked(mid) });
-            base = if cmp == Greater { base } else { mid };
+            if cmp == Equal {
+                return Ok(mid);
+            } else if cmp == Less {
+                base = mid
+            }
             size -= half;
         }
         // SAFETY: base is always in [0, size) because base <= mid.