@nbdd0121 2020-07-10
> Would you be able to put together a benchmark assessing the worst case impact? The new implementation does potentially 50% more conditional branches in the hot loop.

This branch will almost always be predicted correctly.