@nbdd0121 2020-07-10
> I don't know why the docs of binary_search_by said "If there are multiple matches, then any one of the matches could be returned.", but the implementation isn't the same thing. Actually, it returns the last one if multiple matches found.

The doc says any one could be returned precisely because we don't want it to tie to a particular implementation. Returning the last one is obviously also "any one". Changing the implementation to actually return "any one" instead of the last one isn't breaking because that's the contract specified.