@Elinvynia 2020-07-25
Ping from triage:
Hello @Folyd , could you have a look at the review comment by dtolnay when you have the time? Thanks, and let us know if there are any issues!~