@nbdd0121 2020-07-29
So the performance gets slightly worse (only visible with large amount of data) if the value searched has no dups.

BTW: This change needs a crater run as suggested in this test case:
https://github.com/rust-lang/rust/blob/bdbc28a121482d3aac3111482a8167e190f17b46/library/core/tests/slice.rs#L68-L71