@nbdd0121 2020-07-29
I checked the assembly generated, the number of instructions in the hot path increases by 2, where I expect only one. It seems a missed optimisation by LLVM.

@Folyd can you try benchmarking this code sequence instead of the `if/else if`?
```rust
base = if cmp == Greater { base } else { mid };
if cmp == Equal { break }
```