@ranma42 2020-07-29
Quaternary search has been discussed (and rejected) in https://github.com/rust-lang/rust/issues/74547